refactor: move to tsdown (take 2)
