--- .npmignore
@@ -1,18 +1,0 @@
-coverage
-bin
-test
-.*
-*.log
-CONTRIBUTING.md
-tsconfig*.json
-pnpm-workspace.yaml
-eslint.config.js
-*.d.ts
-*.d.ts.map
-!dist/index.d.ts
-!dist/index.d.ts.map
-!dist/getJsdocProcessorPlugin.d.ts
-!dist/getJsdocProcessorPlugin.d.ts.map
-!dist/iterateJsdoc.d.ts
-!dist/iterateJsdoc.d.ts.map
-docs
