+++ package.json
@@ -18,6 +18,7 @@
   },
   "description": "JSDoc linting rules for ESLint.",
   "devDependencies": {
+    "@arethetypeswrong/core": "^0.18.2",
     "@babel/cli": "^7.27.2",
     "@babel/core": "^7.27.4",
     "@babel/eslint-parser": "^7.27.5",
@@ -63,11 +64,12 @@
     "lodash.defaultsdeep": "^4.6.1",
     "mocha": "^11.7.0",
     "open-editor": "^5.1.0",
-    "replace": "^1.2.2",
-    "rimraf": "^6.0.1",
+    "publint": "^0.3.12",
     "semantic-release": "^24.2.5",
+    "tsdown": "^0.13.4",
     "typescript": "5.8.3",
-    "typescript-eslint": "^8.34.1"
+    "typescript-eslint": "^8.34.1",
+    "unplugin-unused": "^0.5.1"
   },
   "engines": {
     "node": ">=20.11.0"
@@ -82,25 +84,42 @@
     "*.js": "eslint --fix"
   },
   "type": "module",
-  "main": "./dist/index.cjs",
-  "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./src/index.js",
-      "require": "./dist/index.cjs"
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "default": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
     },
     "./getJsdocProcessorPlugin.js": {
-      "types": "./dist/getJsdocProcessorPlugin.d.ts",
-      "import": "./dist/getJsdocProcessorPlugin.cjs",
-      "require": "./src/getJsdocProcessorPlugin.js"
+      "import": {
+        "types": "./dist/getJsdocProcessorPlugin.d.ts",
+        "default": "./dist/getJsdocProcessorPlugin.js"
+      },
+      "default": {
+        "types": "./dist/getJsdocProcessorPlugin.d.cts",
+        "default": "./dist/getJsdocProcessorPlugin.cjs"
+      }
     },
     "./iterateJsdoc.js": {
-      "types": "./dist/iterateJsdoc.d.ts",
-      "import": "./dist/iterateJsdoc.cjs",
-      "require": "./src/iterateJsdoc.js"
-    }
+      "import": {
+        "types": "./dist/iterateJsdoc.d.ts",
+        "default": "./dist/iterateJsdoc.js"
+      },
+      "default": {
+        "types": "./dist/iterateJsdoc.d.cts",
+        "default": "./dist/iterateJsdoc.cjs"
+      }
+    },
+    "./package.json": "./package.json"
   },
+  "files": [
+    "dist/"
+  ],
   "name": "eslint-plugin-jsdoc",
   "mocha": {
     "reporter": "dot",
@@ -139,8 +158,7 @@
   },
   "scripts": {
     "tsc": "tsc",
-    "tsc-build": "tsc -p tsconfig-prod.json",
-    "build": "rimraf ./dist && NODE_ENV=production babel ./src --out-file-extension .cjs --out-dir ./dist --copy-files --source-maps --ignore ./src/bin/*.js --no-copy-ignored && replace 'require\\(\"\\.(.*?)\\.[^.]*?\"\\)' 'require(\".$1.cjs\")' 'dist' -r --include=\"*.cjs\" && pnpm tsc-build",
+    "build": "tsdown",
     "check-docs": "babel-node ./src/bin/generateDocs.js --check",
     "create-docs": "pnpm run create-options && babel-node ./src/bin/generateDocs.js",
     "create-rule": "babel-node ./src/bin/generateRule.js",
