+++ test/rules/index.js
@@ -1,4 +1,4 @@
-import config from '../../src/index.js';
+import config from '../../src/plugin.js';
 import camelCase from 'camelcase';
 import {
   ESLint,
