--- tsconfig-prod.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["es2023"],
-    "moduleResolution": "Bundler",
-    "module": "ESNext",
-    "allowJs": true,
-    "checkJs": true,
-    "noEmit": false,
-    "emitDeclarationOnly": true,
-    "declaration": true,
-    "declarationMap": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "target": "es2024",
-    "outDir": "dist"
-  },
-  "include": [
-    "src/**/*.js"
-  ],
-  "exclude": ["node_modules", "src/bin/*.js"]
-}
