+++ tsdown.config.ts
@@ -1,0 +1,32 @@
+import {
+  defineConfig,
+} from 'tsdown';
+
+export default defineConfig({
+  attw: {
+    level: 'error',
+    profile: 'node16',
+  },
+  dts: {
+    sourcemap: true,
+  },
+  entry: [
+    './src/index.js',
+    './src/getJsdocProcessorPlugin.js',
+    './src/iterateJsdoc.js',
+  ],
+  failOnWarn: true,
+  format: [
+    'es',
+    'cjs',
+  ],
+  publint: {
+    level: 'error',
+    strict: true,
+  },
+  sourcemap: true,
+  unbundle: true,
+  unused: {
+    level: 'error',
+  },
+});
